refactor(google-workspace): split resource types and action schemas into their own files
Move the resource type declarations (and the capabilityPermissions /
v1AnnotationsWithPermissions helpers) out of connector.go into resource_types.go,
and relocate the global action schemas next to their handlers in actions.go.
Pure code move with no behavior change; connector.go drops ~280 lines and the
now-unused config import. Mirrors the per-concern file layout used by other
connectors (seismic, docusign).
