+++ pkg/connector/actions.go
@@ -6,6 +6,8 @@ import (
 	"net/mail"
 	"strings"
 
+	config "github.com/conductorone/baton-sdk/pb/c1/config/v1"
+	v2 "github.com/conductorone/baton-sdk/pb/c1/connector/v2"
 	"github.com/conductorone/baton-sdk/pkg/annotations"
 	"github.com/conductorone/baton-sdk/pkg/uhttp"
 	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
@@ -20,6 +22,225 @@ import (
 const (
 	appIdGoogleDocsAndGoogleDrive = int64(55656082996)
 	appIdGoogleCalendar           = int64(435070579839)
+)
+
+// Global (account-level) connector action schemas. Their handlers live in this
+// same file; they are registered in GlobalActions (connector.go).
+var (
+	updateUserStatusActionSchema = &v2.BatonActionSchema{
+		Name: "update_user_status",
+		Arguments: []*config.Field{
+			{
+				Name:        "resource_id",
+				DisplayName: "User Resource ID",
+				Description: "ID of the user resource to update the status of",
+				Field:       &config.Field_StringField{},
+				IsRequired:  true,
+			},
+			{
+				Name:        "is_suspended",
+				DisplayName: "Is Suspended",
+				Description: "Update the user status to suspended or active",
+				Field:       &config.Field_BoolField{},
+				IsRequired:  true,
+			},
+		},
+		ReturnTypes: []*config.Field{
+			{
+				Name:        "success",
+				DisplayName: "Success",
+				Description: "Whether the user resource status was updated successfully",
+				Field:       &config.Field_BoolField{},
+			},
+		},
+		ActionType: []v2.ActionType{v2.ActionType_ACTION_TYPE_ACCOUNT},
+	}
+	transferUserDriveFilesActionSchema = &v2.BatonActionSchema{
+		Name:        "transfer_user_drive_files",
+		DisplayName: "Transfer User Drive Files",
+		Description: "Initiate a Google Drive ownership transfer from one user to another.",
+		Arguments: []*config.Field{
+			{
+				Name:        "resource_id",
+				DisplayName: "Source User Resource ID",
+				Description: "ID of the user resource to transfer Drive ownership from.",
+				Field:       &config.Field_StringField{},
+				IsRequired:  true,
+			},
+			{
+				Name:        "target_resource_id",
+				DisplayName: "Target User Resource ID",
+				Description: "ID of the user resource to receive Drive ownership.",
+				Field:       &config.Field_StringField{},
+				IsRequired:  true,
+			},
+			{
+				Name:        "privacy_levels",
+				DisplayName: "Drive Privacy Levels",
+				Description: "One or more of private, shared. Defaults to both.",
+				Field:       &config.Field_StringSliceField{},
+				IsRequired:  false,
+			},
+		},
+		ReturnTypes: []*config.Field{
+			{
+				Name:        "success",
+				DisplayName: "Success",
+				Description: "Whether the transfer request was created successfully.",
+				Field:       &config.Field_BoolField{},
+			},
+			{
+				Name:        "transfer_id",
+				DisplayName: "Transfer ID",
+				Description: "ID of the Data Transfer request.",
+				Field:       &config.Field_StringField{},
+			},
+			{
+				Name:        "status",
+				DisplayName: "Transfer Status",
+				Description: "Initial status returned by the Data Transfer API (e.g., IN_PROGRESS).",
+				Field:       &config.Field_StringField{},
+			},
+		},
+		ActionType: []v2.ActionType{v2.ActionType_ACTION_TYPE_ACCOUNT},
+	}
+	transferUserCalendarActionSchema = &v2.BatonActionSchema{
+		Name:        "transfer_user_calendar",
+		DisplayName: "Transfer User Calendar",
+		Description: "Initiate a Google Calendar transfer from one user to another.",
+		Arguments: []*config.Field{
+			{
+				Name:        "resource_id",
+				DisplayName: "Source User Resource ID",
+				Description: "ID of the user resource to transfer calendar data from.",
+				Field:       &config.Field_StringField{},
+				IsRequired:  true,
+			},
+			{
+				Name:        "target_resource_id",
+				DisplayName: "Target User Resource ID",
+				Description: "ID of the user resource to receive calendar data.",
+				Field:       &config.Field_StringField{},
+				IsRequired:  true,
+			},
+			{
+				Name:        "release_resources",
+				DisplayName: "Release Resources",
+				Description: "If true, sets RELEASE_RESOURCES=TRUE (release resources for future events).",
+				Field:       &config.Field_BoolField{},
+				IsRequired:  false,
+			},
+		},
+		ReturnTypes: []*config.Field{
+			{
+				Name:        "success",
+				DisplayName: "Success",
+				Description: "Whether the transfer request was created successfully.",
+				Field:       &config.Field_BoolField{},
+			},
+			{
+				Name:        "transfer_id",
+				DisplayName: "Transfer ID",
+				Description: "ID of the Data Transfer request.",
+				Field:       &config.Field_StringField{},
+			},
+			{
+				Name:        "status",
+				DisplayName: "Transfer Status",
+				Description: "Initial status returned by the Data Transfer API (e.g., IN_PROGRESS).",
+				Field:       &config.Field_StringField{},
+			},
+		},
+		ActionType: []v2.ActionType{v2.ActionType_ACTION_TYPE_ACCOUNT},
+	}
+	changeUserPrimaryEmailActionSchema = &v2.BatonActionSchema{
+		Name:        "change_user_primary_email",
+		DisplayName: "Change User Primary Email",
+		Description: "Update a user's primary email address.",
+		Arguments: []*config.Field{
+			{
+				Name:        "resource_id",
+				DisplayName: "User Resource ID",
+				Description: "ID of the user resource to update.",
+				Field:       &config.Field_StringField{},
+				IsRequired:  true,
+			},
+			{
+				Name:        "new_primary_email",
+				DisplayName: "New Primary Email",
+				Description: "New primary email address (must be within a verified domain).",
+				Field:       &config.Field_StringField{},
+				IsRequired:  true,
+			},
+		},
+		ReturnTypes: []*config.Field{
+			{
+				Name:        "success",
+				DisplayName: "Success",
+				Description: "Whether the primary email was updated successfully.",
+				Field:       &config.Field_BoolField{},
+			},
+			{
+				Name:        "previous_primary_email",
+				DisplayName: "Previous Primary Email",
+				Description: "User's previous primary email address.",
+				Field:       &config.Field_StringField{},
+			},
+			{
+				Name:        "new_primary_email",
+				DisplayName: "New Primary Email",
+				Description: "User's updated primary email address.",
+				Field:       &config.Field_StringField{},
+			},
+		},
+		ActionType: []v2.ActionType{v2.ActionType_ACTION_TYPE_ACCOUNT},
+	}
+	disableUserActionSchema = &v2.BatonActionSchema{
+		Name:        "disable_user",
+		DisplayName: "Disable User",
+		Description: "Suspend a user account.",
+		Arguments: []*config.Field{
+			{
+				Name:        "user_id",
+				DisplayName: "User Resource ID",
+				Description: "ID of the user resource to disable (suspend).",
+				Field:       &config.Field_StringField{},
+				IsRequired:  true,
+			},
+		},
+		ReturnTypes: []*config.Field{
+			{
+				Name:        "success",
+				DisplayName: "Success",
+				Description: "Whether the user was disabled (suspended) successfully.",
+				Field:       &config.Field_BoolField{},
+			},
+		},
+		ActionType: []v2.ActionType{v2.ActionType_ACTION_TYPE_ACCOUNT_DISABLE},
+	}
+	enableUserActionSchema = &v2.BatonActionSchema{
+		Name:        "enable_user",
+		DisplayName: "Enable User",
+		Description: "Unsuspend a user account.",
+		Arguments: []*config.Field{
+			{
+				Name:        "user_id",
+				DisplayName: "User Resource ID",
+				Description: "ID of the user resource to enable (unsuspend).",
+				Field:       &config.Field_StringField{},
+				IsRequired:  true,
+			},
+		},
+		ReturnTypes: []*config.Field{
+			{
+				Name:        "success",
+				DisplayName: "Success",
+				Description: "Whether the user was enabled (unsuspended) successfully.",
+				Field:       &config.Field_BoolField{},
+			},
+		},
+		ActionType: []v2.ActionType{v2.ActionType_ACTION_TYPE_ACCOUNT_ENABLE},
+	}
 )
 
 func (c *GoogleWorkspace) updateUserStatus(ctx context.Context, args *structpb.Struct) (*structpb.Struct, annotations.Annotations, error) {
